+++ ts/packages/core/src/lib/toolRouterParams.ts
@@ -1,0 +1,74 @@
+import { SessionCreateParams } from '@composio/client/resources/tool-router.mjs';
+import {
+  ToolRouterConfigTags,
+  ToolRouterConfigTools,
+  ToolRouterConfigToolsSchema,
+  ToolRouterToolsParam,
+} from '../types/toolRouter.types';
+import { ValidationError } from '../errors';
+
+export const transformToolRouterToolsParams = (
+  params?: Record<string, ToolRouterToolsParam | ToolRouterConfigTools> | undefined
+):
+  | Record<
+      string,
+      SessionCreateParams.Enable | SessionCreateParams.Disable | SessionCreateParams.Tags
+    >
+  | undefined => {
+  if (!params) {
+    return undefined;
+  }
+
+  if (typeof params === 'object') {
+    const result = Object.keys(params).reduce(
+      (acc, key) => {
+        if (Array.isArray(params[key])) {
+          acc[key] = { enable: params[key] };
+        } else if (typeof params[key] === 'object') {
+          const parsedResult = ToolRouterConfigToolsSchema.safeParse(params[key]);
+          if (parsedResult.success) {
+            const data = parsedResult.data;
+            if (Array.isArray(data)) {
+              acc[key] = { enable: data };
+            } else if ('enable' in data) {
+              acc[key] = { enable: data.enable };
+            } else if ('disable' in data) {
+              acc[key] = { disable: data.disable };
+            } else if ('tags' in data) {
+              const tags = transformToolRouterTagsParams(data.tags);
+              if (tags) {
+                acc[key] = { tags };
+              }
+            }
+          } else {
+            throw new ValidationError(parsedResult.error.message);
+          }
+        } else {
+          acc[key] = { enable: params[key] };
+        }
+        return acc;
+      },
+      {} as Record<
+        string,
+        SessionCreateParams.Enable | SessionCreateParams.Disable | SessionCreateParams.Tags
+      >
+    );
+    return result;
+  }
+};
+
+export const transformToolRouterTagsParams = (
+  params?: ToolRouterConfigTags
+): SessionCreateParams.Tags['tags'] | undefined => {
+  if (!params) {
+    return undefined;
+  }
+  if (Array.isArray(params)) {
+    return { enable: params };
+  } else if (typeof params === 'object') {
+    return {
+      enable: params.enable,
+      disable: params.disable,
+    };
+  }
+};